api/requirements/requirement.txt
# Django
Django==2.2.7
django-environ==0.4.5
django-redis==4.12.1
requests==2.23.0

# Python-PostgreSQL Database Adapter
psycopg2-binary==2.8.4

# Models
django-model-utils==3.2.0
python_dateutil==2.8.0

# DRF
djangorestframework==3.10.3
django-oauth-toolkit==1.2.0
django-cors-middleware==1.4.0
django-rest-auth==0.9.5
django-allauth==0.40.0


# Static files
# whitenoise==4.1.4  # https://github.com/evansd/whitenoise


# Document
drf-yasg==1.17.0

# debug
django_extensions==3.1.3
django-debug-toolbar==3.2.1

# WSGI Handler
gunicorn==20.0.2


# JWT
PyJWT==2.1.0
djangorestframework_simplejwt==4.7.2
cryptography==3.4.7

# bdist_wheel
wheel==0.36.2

# Image
Pillow==7.0.0

# Testing
coverage==4.5.4


# Fake data
Faker==2.0.4
